+++ changelog.md
@@ -1,5 +1,7 @@
 ## Upcoming
 * Added back Farming features: Crops, Bone Meal
+  * Rich farmland has been disabled by default
+  * Crops can now grow with bone meal when farmland is not moist
 * Fixed Water Fall Damage feature not working properly
 
 ## 5.2.0.0-alpha